Make chip-layout generic
