+++ app/components/ChipCollection/styles.css
@@ -1,0 +1,24 @@
+.chip-component {
+    display: flex;
+    align-items: flex-end;
+    flex-wrap: wrap;
+    padding: var(--dui-spacing-large);
+    gap: var(--dui-spacing-large);
+
+    .chip-filter-header {
+        color: var(--color);
+        font-size: var(--dui-font-size-small);
+        font-weight: var(--dui-font-weight-bold);
+    }
+
+    .chip-collection {
+        display: flex;
+        flex-wrap: wrap;
+        gap: var(--dui-spacing-medium);
+
+        .chip-action-buttons {
+            padding: var(--dui-spacing-extra-small);
+        }
+
+    }
+}
